Fixed threading issues with the Transfer operation
The transfer operation was attempting to reuse the SCU connection to the target, which was causing conflicts with the presentation context. It was also using the MuleSoft scheduled executor for the SCU connection, which was sometimes causing those transactions to be scheduled outside of the transfer window. Now the transfer operation creates its own single-thread executor for the SCU connections, effectively serializing the C-STORE for each DICOM file received from the C-GET.